+++ docs/changes.rst
@@ -16,6 +16,8 @@ To be released.
 - Fixed a bug that raised :exc:`~urllib.error.HTTPError` when
   non-existent `Entity` was requested.  [:pr:`11`]
 - Added support for `time` datatypes with precision `9` (year-only).  [:pr:`26`]
+- Added support for decoding the ``globe-coordinate`` datatype.  [:pr:`28`]
+
 
 Version 0.6.1
 -------------